ai_assistant/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
aiofiles==0.7.0
python-dotenv==0.19.0
openai==0.27.0
gitpython==3.1.24
chromadb==0.3.21
sentence-transformers==2.2.2
pytest==6.2.5
pytest-asyncio==0.15.1
black==21.9b0
isort==5.9.3
flake8==3.9.2